requirements.txt
SpeechRecognition
pyttsx3
scikit-learn
numpy
tensorflow
keras
torch
transformers
